requirements.txt
flask
python-dateutil
openai
